+++ kafka-monitor/settings.py
@@ -4,7 +4,7 @@ REDIS_PORT = 6379
 
 # Kafka server information
 KAFKA_HOSTS = ''
-KAFKA_INCOMING_TOPIC = 'demo.incoming_urls'
+KAFKA_INCOMING_TOPIC = 'demo.incoming'
 KAFKA_GROUP = 'demo-group'
 
 PLUGINS = {}